pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dumprx"
version = "2.0.0"
description = "Advanced firmware dumper and extractor toolkit"
authors = [{name = "DumprX Team"}]
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "aiohttp>=3.8.0",
    "aiofiles>=23.0.0",
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "gitpython>=3.1.0",
    "python-telegram-bot>=20.0",
    "humanize>=4.0.0",
    "cryptography>=3.4.0",
    "py7zr>=0.20.0",
    "rarfile>=4.0",
    "patool>=1.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
dumprx = "dumprx.cli:main"

[project.urls]
Homepage = "https://github.com/Eduardob3677/DumprX"
Repository = "https://github.com/Eduardob3677/DumprX"
Issues = "https://github.com/Eduardob3677/DumprX/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["dumprx*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true